removed .pogo extension from requires
This is because of a strange bug: never mind the bug itself, but
the cause was because we had two requires in the source:

    require 'x.pogo'

And

    require 'x'

The file `x.pogo` had been compiled to `x.js` too, so these
two requires resulted in requiring two different files,
`x.pogo` and `x.js` which one would expect. However, because
they had the same source code, the code was effectively run twice.

Better to always just omit the `.pogo` extension.